Rust/His: Add fast byte scanner for HISv2
We need a way to quickly scan for highly identifiable secrets in both
files and raw byte streams of unknown encodings. Due to the way HISv2
works, we can scan for the signature JQQJ and know if the data is UTF8
or UTF16.

To enable this quickly, we first scan 16 bytes at a time for the letter
J via loop that LLVM vectorizes. If the letter J is found, then we start
acccumulating bytes into a 64-bit value. 8 bytes is enough to detect
JQQJ in UTF8 and UTF16. We can check all bytes in a single instruction
vs having to check one by one. We only have to compare the accumulator
once we add J to the accumulator.

Once we detect JQQJ, we save the location the detection for later
further full analysis. This allows for the scanner to scan large files
and only save the possible locations to check. It can also scan small or
large byte buffers.

The possible matches are fetched from memory or a file, then they are
sent through a full validation of the data. The full validation so far
is only the pattern/regex matching. It does not currently include the
marvin checksum, etc.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>