Rust/His: Fix typo and add file unit tests
There was a typo in the matching of an equals sign, which this fixes.

Add unit tests for file readings, which uncovered a corner case. We
cannot assume explicit sizes of bytes, since files may be truncated. We
must move to a range of bytes when determining how to parse the bytes.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>