Scans (#10)
* Rust/His: Add fast byte scanner for HISv2

We need a way to quickly scan for highly identifiable secrets in both
files and raw byte streams of unknown encodings. Due to the way HISv2
works, we can scan for the signature JQQJ and know if the data is UTF8
or UTF16.

To enable this quickly, we first scan 16 bytes at a time for the letter
J via loop that LLVM vectorizes. If the letter J is found, then we start
acccumulating bytes into a 64-bit value. 8 bytes is enough to detect
JQQJ in UTF8 and UTF16. We can check all bytes in a single instruction
vs having to check one by one. We only have to compare the accumulator
once we add J to the accumulator.

Once we detect JQQJ, we save the location the detection for later
further full analysis. This allows for the scanner to scan large files
and only save the possible locations to check. It can also scan small or
large byte buffers.

The possible matches are fetched from memory or a file, then they are
sent through a full validation of the data. The full validation so far
is only the pattern/regex matching. It does not currently include the
marvin checksum, etc.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

* Rust/His: Fix typo and add file unit tests

There was a typo in the matching of an equals sign, which this fixes.

Add unit tests for file readings, which uncovered a corner case. We
cannot assume explicit sizes of bytes, since files may be truncated. We
must move to a range of bytes when determining how to parse the bytes.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

* Rust/His: Add HisV1 32 and 64 byte signatures

We have various HisV1 4 byte signatures, each indicates the byte length
of the keys. Many of these are 32 and 64 byte signatures.

Add initial support for scanning for only v2, v1, or both.
Add unit test to ensure known signatures match for 32 and 64 byte V1.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

* His/Rust: Fix bug and ensure ScanMatchType is ABI

ScanMatchType will be exposed out via FFI to other languages. We need to
ensure these values do not change.

Define constants to ScanMatchType to ensure they don't change.
Ensure we always check byte range before checking patterns.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

* His/Rust: Expose Scan and ScanOptions for FFI

Scan and ScanOptions are behind a non-public module.

Ensure we have both Scan and ScanOptions exported for callers to
utilize.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

* His/Rust: Make FFI compatible crate

Add an FFI compatible crate that can be called from C like languages.

This crate just links in the normal core security crate and wraps it up
with C exports.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

* His/DotNet: Utilize the FFI core crate for scanning

Add a new IdentifiableScan class in the core DLL that utilizes the FFI
crate dylib.

Update the CLI tool to have a new "scan" command that uses the new
IdentifiableScan class to scan for identifiable secrets within a given
file.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>

---------

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>
Co-authored-by: Beau Belgrave <beaub@linux.microsoft.com>