Rust/His: Fix trailing UTF16 case
For UTF16 signature cases, we can have trailing data that is not an
equals sign. We need to always used the MATCHED length back when giving
back UTF8 text. This removes the trailing data unless it's an equal
sign.

Use len vs count for UTF8 byte len when giving back match results.
Update unit test to ensure this case works.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>