Rust/His: Add interop that allows specifying definitions
We don't always need to scan for everything. We often might only need to
scan for certain rules.

Update the interop in Rust to accomodate this by allowing a filter
string to be passed with a semi-colon delimited list of IDs to use.

Update the interop in C# to utilize this via the RegexPatterns that are
passed. Only the patterns that are passed are enabled.

Signed-off-by: Beau Belgrave <beaub@linux.microsoft.com>