feat(identifiable_scans)!: don't require allocation for valid utf8 data in ScanMatch
In the case where we are processing utf8 data, there is no need to put that data
on the heap. We can simply pass a `Cow<str>` pointer that contains the data. In the
utf16 case, we can have the `Cow<str>` own the transformed data.

BREAKING: changes the type signature of ScanMatch to include the lifetime
of the matched text.