Rust/c3id: Decrease allocations & formatting (#74)
* fix(cross_company_correlating_id): Decrease allocations & formatting calls during generation of c3id

There are some steps in the generation of the Cross Company Correlation ID
for which we know the amount and type of data that we will be processing at
compile time. Passing this knowledge directly to those functions decreases
the amount of time spent formatting & reallocating at runtime.

* fix(cross_company_correlating_id): remove final allocation

Thanks to a comment by @suvamM I realized that there is no need to
allocate the string data at all. Since we know the length of the data
up front, and the formatting we do is so basic that doing it by hand
works fine, we can remove the allocation and gain another 7% speedup.

* fix(cross_company_correlating_id): borrow only once

* fix(cross_company_correlating_id): document hex encoding