Run tests on .NET 8, fix issues uncovered by that (#122)
This started with wanting to examine `IdentifiableScan` performance. I profiled its unit test and the profiler told me that it was spending a lot of time in regular expressions and so I probably wanted to use a newer version of .NET. :) This made me realize that the tests were running only on .NET Framework. I enabled them on .NET 8 and when I ran them, I was greeted with `System.NotSupportedException: The specified pattern with RegexOptions.NonBacktracking could result in an automata as large as 'NNNN' nodes, which is larger than the configured limit of '1000'.`

This was due to `ValidateBase64Key` generating a regex with a large quantifier for large input. I fixed this by removing the regex generation and use. Instead, we only check that 1) there is not inconsistent use of base64url or standard base64 with respect to the `encodeForUrl` argument, and 2) the key has the signature provided. The remainder of the work done by the generated regex was redundant to what had already been ascertained by `ValidateChecksum`. In addition to fixing the bug, this
improved the performance of the slowest test, `IdentifiableSecrets_GenerateBase64Key_Comprehensive`, by about 8x on .NET Framework.

While looking at this, I further discovered that `ValidateBase64Key`  did not check the signature if given a backwards-compatible `CommonAnnotatedKey`, and that `ValidateCommonAnnotatedKey` didn't check the signature at all. Also, the `byte[]` version of ValidateCommonAnnotatedKey did not validate the signature argument. These bugs are fixed here as well.