Fix incorrect signatures preventing rust scanner from being used for two patterns.
`IdentifiableScan` falls back to the slower `SecretMasker` for patterns that aren't detected by rust. Determining if rust can detect a pattern is controlled by a hard-coded list of signatures. In two cases, HIS v2 and Azure IOT, the values in the hard-coded list did not match the signatures of their `RegexPattern`s. This results in using the backup slower masker for these two patterns.

Fixing this causes tests to fail when rust fails to match `JQQJ9D` (derived) and `JQQJ9H` (hashed) patterns. This draft adds support for these to rust, but these formats never got off the ground and should probably be removed throughout instead.

Fixing this also reveals that the C# masker was including the non-alphanumeric character preceding an HIS v2 secret in its match. Once `IdentifiableScan` started to use rust for HIS v2 matching, tests failed because the C3IDs didn't match. This is fixed as well by using "refine" as the capture name instead of "secret".