Sort properties in exported JSON
Currently, property order is influenced by whether or not a property is overridden or introduced by a derived class. This change orders the properties by name irrespective of where they appear in the class hierarchy.

Submitting it on its own so that a pending change that removes several overrides will be easier to review.