Allow default regex options like NonBacktracking to be disabled.
In the azure-pipelines-agent scenario, we have to support arbitrary .NET regexes, which may be incompatible with non-backtracking. However, there was no way to stop the library from using it when running on .NET 8+. This change provides a way to specify during RegexPattern construction that certain default regex options should be disabled for that pattern. To support this, the regex engine itself no longer adds the defaults and it is left to the caller. Engine callers may still omit the options argument to get the defaults, but if they pass options, then the options passed, and only the options passed, will be used.

This involved some refactoring as the regex option defaults were spread in multiple places. They are now consolidated to one place: `RegexDefaults.DefaultOptions` (renamed from `RegexDefaults.DefaultOptionsCaseSensitive). This consolidation revealed that we were not in fact using `RegexOptions.CultureInvariant` even though it was specified there. Now we use this unless it is removed using the new feature described above.