Allow default regex options like NonBacktracking to be disabled (#185)
In the azure-pipelines-agent scenario, we have to support arbitrary .NET regexes, which may be incompatible with non-backtracking. However, there was no way to stop the library from using it when running on .NET 8+. 

This change makes any explicit `RegexOptions` passed to `RegexPattern` constructor or `IRegexEngine`, the definitive full set of options to use. If callers want the default plus some additional options, it is now their responsibility to `|` their additional options onto `RegexDefaults.DefaultOptions`.

`RegexOptions` default arguments are now nullable and null by default. Null indicates "use the default options". Nullable is used to prevent any future change to the defaults from requiring callers to be recompiled. It also deals with a corner case where a portable library could compile against netstandard2.0 and run on net8.0+ and lose `NonBacktracking`.

This involved some refactoring as the regex option defaults were spread in multiple places. They are now consolidated to one place: `RegexDefaults.DefaultOptions` (renamed from `RegexDefaults.DefaultOptionsCaseSensitive`). This consolidation revealed that we were not in fact using `CultureInvariant` even though it was specified there. Now we actually use it by default everywhere.