Do not send duplicate detections to MaskSecrets callback (#216)
While we still send detections that are nested in other detections, partially overlap other detections, or are adjacent to other detections, this change no longer sends more than one detection with the exact same start position and length. The main motivation here is for the use case where there are literal values added to the masker that also match regexes. The literal value is a known secret and may commonly also match a regex that finds secrets of its kind. The regex match is secondary in this case and not something that the caller will want to act upon.

This change also improves the reliability of preferring literal redaction token over regex redaction token by not having it rely on `"***"` (default literal) sorting before `"+++"` (default regex) and `"SECNNN/NNN:C3ID"` (high entropy rule regex with generateCorrelatingIds = true). The default `"***"` and `"+++"` tokens could be customized such that this would stop working. The sort order now explicitly prioritizes literal matches over regex matches, irrespective of redaction tokens. We also sort by ID before sorting by redaction token so that the detections sent to the callback do not change depending on generateCorrelatingIds = true or false.