Add DetectSecrets and MaskSecrets overloads that take `ReadOnlyMemory<char>`
Internally, this is implemented using a `StringInput` struct that is backed by ReadOnlyMemory<char> on modern .NET and string on .NET Framework/Standard.

This is supported only on modern .NET 8+. Custom regex engines will throw if used with this API unless they are updated to implement a new Matches overload that takes Span<char>.

RegexPattern.GetDetections and SecretLiteral are now internal. These API needed breaking changes to implement this and it will be better to keep them as implementation details moving forward.