Remove `identifiable_secrets` module from rust crate (#234)
All indications are that no rust consumers are using `identifiable_secrets` which provides secret *allocation* and *masking*. Furthermore, it has severe issues that are not worth fixing if there are no consumers.

Note that `identifiable_scans` (for secret *scanning*), `marvin`, and `cross_company_correlating_id` remain and are unchanged as these are actually used.

Tests started failing in August 2025, blocking all PRs. This is the same issue as #90, but there was another test that used capital 'A' rather than 'a' as the test_char and escaped until now.

Trying the tests with real random keys reveals severe bugs that manifest as failing ~5% of runs on my machine.
- `SecretMasker` assumes all detections are `SEC101/200`.
- `SecretMasker` assumes all `PossibleScanMatch` are definite matches.

If we need to bring this code back, the above bugs should be fixed and the follow-up noted in #90 should be done at that time.

Also in this PR:
- Add release notes for prior changes that neglected them.
- Remove some commented out code that led to false matches while grep'ing during the investigation.